undid the Arguments.checkInteger change since it can break things (and did!), checking for actual int byte mask in hologram manually now

+++ src/main/scala/li/cil/oc/server/network/Component.scala
@@ -259,7 +259,7 @@ object Component {
     def checkInteger(index: Int) = {
       checkIndex(index, "number")
       args(index) match {
-        case value: java.lang.Double => value.longValue.intValue
+        case value: java.lang.Double => value.intValue
         case value => throw typeError(index, value, "number")
       }
     }
